Fixed blurry image on detection + added resizing of the window
Fixed glitch that would make the image go blurry on detection activated
also now you can resize the windows of the cameras

+++ src/main/java/test/camera/SerializedCamera.java
@@ -16,8 +16,8 @@ public class SerializedCamera implements Serializable {
 	public int x = 20;
 	public int y = 20;
 
-	public int height = 400;
-	public int width = 400;
+	public double height = 400;
+	public double width = 400;
 
 	public boolean motionDetection = false;
 	public int threshold = 15; //max = 255
